[FIX] 0037930: Race Condition in Unzipping

+++ Services/Utilities/classes/class.ilUtil.php
@@ -1588,95 +1588,50 @@ class ilUtil
         //@chmod($a_dir, $a_mod);
     }
 
-
-    /**
-    * unzip file
-    *
-    * @param	string	$a_file		full path/filename
-    * @param	boolean	$overwrite	pass true to overwrite existing files
-    * @static
-    *
-    */
-    public static function unzip($a_file, $overwrite = false, $a_flat = false)
-    {
+    public static function unzip(
+        string $path_to_zip_file,
+        bool $overwrite_existing = false,
+        bool $unpack_flat = false
+    ) {
         global $DIC;
 
         $log = $DIC->logger()->root();
 
-        if (!is_file($a_file)) {
+        if (!is_file($path_to_zip_file)) {
             return;
         }
 
-        // if flat, move file to temp directory first
-        if ($a_flat) {
-            $tmpdir = ilUtil::ilTempnam();
-            ilUtil::makeDir($tmpdir);
-            copy($a_file, $tmpdir . DIRECTORY_SEPARATOR . basename($a_file));
-            $orig_file = $a_file;
-            $a_file = $tmpdir . DIRECTORY_SEPARATOR . basename($a_file);
-            $origpathinfo = pathinfo($orig_file);
-        }
+        // we unpack the zip always in a temp directory
+        $temporary_unzip_directory = ilUtil::ilTempnam();
+        ilUtil::makeDir($temporary_unzip_directory);
+        copy($path_to_zip_file, $temporary_unzip_directory . DIRECTORY_SEPARATOR . basename($path_to_zip_file));
+        $original_path_to_zip_file = $path_to_zip_file;
+        $path_to_zip_file = $temporary_unzip_directory . DIRECTORY_SEPARATOR . basename($path_to_zip_file);
+        $original_zip_path_info = pathinfo($original_path_to_zip_file);
+        $unzippable_zip_path_info = pathinfo($path_to_zip_file);
 
-        $pathinfo = pathinfo($a_file);
-        $dir = $pathinfo["dirname"];
-        $file = $pathinfo["basename"];
+        $unzippable_zip_directory = $unzippable_zip_path_info["dirname"];
+        $unzippable_zip_filename = $unzippable_zip_path_info["basename"];
 
         // unzip
-        $cdir = getcwd();
-        chdir($dir);
-        $unzip = PATH_TO_UNZIP;
-
-        // the following workaround has been removed due to bug
-        // http://www.ilias.de/mantis/view.php?id=7578
-        // since the workaround is quite old, it may not be necessary
-        // anymore, alex 9 Oct 2012
-        /*
-                // workaround for unzip problem (unzip of subdirectories fails, so
-                // we create the subdirectories ourselves first)
-                // get list
-                $unzipcmd = "-Z -1 ".ilUtil::escapeShellArg($file);
-                $arr = ilUtil::execQuoted($unzip, $unzipcmd);
-                $zdirs = array();
-
-                foreach($arr as $line)
-                {
-                    if(is_int(strpos($line, "/")))
-                    {
-                        $zdir = substr($line, 0, strrpos($line, "/"));
-                        $nr = substr_count($zdir, "/");
-                        //echo $zdir." ".$nr."<br>";
-                        while ($zdir != "")
-                        {
-                            $nr = substr_count($zdir, "/");
-                            $zdirs[$zdir] = $nr;				// collect directories
-                            //echo $dir." ".$nr."<br>";
-                            $zdir = substr($zdir, 0, strrpos($zdir, "/"));
-                        }
-                    }
-                }
-
-                asort($zdirs);
-
-                foreach($zdirs as $zdir => $nr)				// create directories
-                {
-                    ilUtil::createDirectory($zdir);
-                }
-        */
+        $current_directory = getcwd();
+        chdir($unzippable_zip_directory);
+        $unzip_command = PATH_TO_UNZIP;
 
         // real unzip
-        if (!$overwrite) {
-            $unzipcmd = ilUtil::escapeShellArg($file);
+        if (!$overwrite_existing) {
+            $unzip_parameters = ilUtil::escapeShellArg($unzippable_zip_filename);
         } else {
-            $unzipcmd = "-o " . ilUtil::escapeShellArg($file);
+            $unzip_parameters = "-o " . ilUtil::escapeShellArg($unzippable_zip_filename);
         }
-        ilUtil::execQuoted($unzip, $unzipcmd);
-
-        chdir($cdir);
+        ilUtil::execQuoted($unzip_command, $unzip_parameters);
+        // move back
+        chdir($current_directory);
 
         // remove all sym links
         clearstatcache();			// prevent is_link from using cache
-        $dir_realpath = realpath($dir);
-        foreach (new RecursiveIteratorIterator(new RecursiveDirectoryIterator($dir)) as $name => $f) {
+        $dir_realpath = realpath($unzippable_zip_directory);
+        foreach (new RecursiveIteratorIterator(new RecursiveDirectoryIterator($unzippable_zip_directory)) as $name => $f) {
             if (is_link($name)) {
                 $target = readlink($name);
                 if (substr($target, 0, strlen($dir_realpath)) != $dir_realpath) {
@@ -1686,20 +1641,31 @@ class ilUtil
             }
         }
 
-        // if flat, get all files and move them to original directory
-        if ($a_flat) {
-            include_once("./Services/Utilities/classes/class.ilFileUtils.php");
-            $filearray = array();
-            ilFileUtils::recursive_dirscan($tmpdir, $filearray);
-            if (is_array($filearray["file"])) {
-                foreach ($filearray["file"] as $k => $f) {
-                    if (substr($f, 0, 1) != "." && $f != basename($orig_file)) {
-                        copy($filearray["path"][$k] . $f, $origpathinfo["dirname"] . DIRECTORY_SEPARATOR . $f);
+        // rename executables
+        self::renameExecutables($unzippable_zip_directory);
+
+        // now we have to move the files to the original directory. if $a_flat is true, we move the files only without directories, otherwise we move the whole directory
+        if ($unpack_flat) {
+            $file_array = [];
+            ilFileUtils::recursive_dirscan($temporary_unzip_directory, $file_array);
+            if (is_array($file_array["file"])) {
+                foreach ($file_array["file"] as $k => $f) {
+                    if (
+                        substr($f, 0, 1) !== "."
+                        && $f !== basename($original_path_to_zip_file)
+                    ) {
+                        copy(
+                            $file_array["path"][$k] . $f,
+                            $original_zip_path_info["dirname"] . DIRECTORY_SEPARATOR . $f
+                        );
                     }
                 }
             }
-            ilUtil::delDir($tmpdir);
+        } else {
+            self::rCopy($temporary_unzip_directory, $original_zip_path_info["dirname"]);
         }
+
+        ilUtil::delDir($temporary_unzip_directory);
     }
 
     /**
@@ -3584,10 +3550,12 @@ class ilUtil
     * @static
     *
     */
-    public static function rRenameSuffix($a_dir, $a_old_suffix, $a_new_suffix)
+    public static function rRenameSuffix(string $a_dir, string $a_old_suffix, string $a_new_suffix) : bool
     {
-        if ($a_dir == "/" || $a_dir == "" || is_int(strpos($a_dir, ".."))
-            || trim($a_old_suffix) == "") {
+        if ($a_dir === "/"
+            || $a_dir === ""
+            || strpos($a_dir, "..") !== false
+            || trim($a_old_suffix) === "") {
             return false;
         }
 
@@ -3598,19 +3566,28 @@ class ilUtil
 
         // read a_dir
         $dir = opendir($a_dir);
+        if ($dir === false) {
+            return false;
+        }
+
+        $prohibited =  [
+            '...'
+        ];
 
         while ($file = readdir($dir)) {
-            if ($file != "." and
-            $file != "..") {
+            if (
+                $file !== "."
+                && $file !== ".."
+            ) {
                 // triple dot is not allowed in filenames
-                if ($file === '...') {
+                if (in_array($file, $prohibited)) {
                     unlink($a_dir . "/" . $file);
                     continue;
                 }
 
                 // directories
                 if (@is_dir($a_dir . "/" . $file)) {
-                    ilUtil::rRenameSuffix($a_dir . "/" . $file, $a_old_suffix, $a_new_suffix);
+                    self::rRenameSuffix($a_dir . "/" . $file, $a_old_suffix, $a_new_suffix);
                 }
 
                 // files
@@ -3630,8 +3607,7 @@ class ilUtil
 
                     $path_info = pathinfo($a_dir . "/" . $file);
 
-                    if (strtolower($path_info["extension"]) ==
-                    strtolower($a_old_suffix)) {
+                    if (strtolower($path_info["extension"]) === strtolower($a_old_suffix)) {
                         $pos = strrpos($a_dir . "/" . $file, ".");
                         $new_name = substr($a_dir . "/" . $file, 0, $pos) . "." . $a_new_suffix;
                         // check if file exists
